requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0               # PostgreSQL async driver
alembic==1.14.0               # DB migrations

# ── Configuration & Validation ─────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3       # EmailStr support

# ── Security ───────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt<4.0                        # passlib 1.7.4 incompatible with bcrypt>=4.0

# ── HTTP Client (external billing system) ─────────────────────────────────────
httpx==0.27.2

# ── Logging ────────────────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Rate limiting ──────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── PDF generation ─────────────────────────────────────────────────────────────
# WeasyPrint is the primary engine (Linux/Docker); ReportLab is the Windows fallback.
# Both are installed so the correct one is used at runtime automatically.
weasyprint==62.3                  # Primary: HTML/CSS → PDF (requires system GTK/Cairo/Pango)
reportlab==4.2.5                  # Fallback: pure-Python PDF generation (Windows-compatible)
jinja2==3.1.4                     # HTML template rendering (used by WeasyPrint path)

# ── Email delivery ─────────────────────────────────────────────────────────────
resend==2.4.0                     # Official Resend Python SDK

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0             # In-memory SQLite for tests (no Postgres needed)
